% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors_Dev.R
\docType{data}
\name{rpg_color_list}
\alias{rpg_color_list}
\title{Full Color List}
\format{
An object of class \code{character} of length 71.
}
\usage{
rpg_color_list
}
\description{
A full list of the individual colors used in this package, including name and
hex code
}
\keyword{datasets}
