% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors_Dev.R
\name{rpg_color_pal}
\alias{rpg_color_pal}
\title{Retrieve a Palette}
\usage{
rpg_color_pal(palette = "logo", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A string representing a color palette, see https://bfroebrpg.github.io/rpgcolorsr/
for a full list of palettes}

\item{reverse}{When true the default order of the color palette is reversed}
}
\value{
a function to create a color palette
}
\description{
This function excepts a character string for a palette and returns a function
that can be turned into a palette
}
\details{
adding a number after the function call will create a palette with that number
of steps
}
