% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors_Dev.R
\name{rpg_cols}
\alias{rpg_cols}
\title{Retrieve a Color}
\usage{
rpg_cols(...)
}
\description{
This function takes a string or group of strings and returns hex codes
for the corresponding colors
}
\details{
A full list of the colors in this package can be returned by running the
function with no input
}
