% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colors_Dev.R
\name{scale_fill_rpg}
\alias{scale_fill_rpg}
\title{Create a Fill Scale for ggplot}
\usage{
scale_fill_rpg(palette = "logo", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{A string representing a color palette, see the insert github
link to markdown doc to get a full list of palettes}

\item{discrete}{Changes between discrete or continuous scales}

\item{reverse}{When true the default order of the color palette is reversed}
}
\value{
a function to create a color palette
}
\description{
This function is inserted into a ggplot call to replace a scale_fill
command.
}
\details{
By default the palette will be discrete.
}
